+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.2] - 2020-7-14
+
+ - Added support for map category and layout properties.
+ - Backend refactoring to make the editor UI far more extensible.
+
 ## [1.0.1] - 2020-4-29
 
 This release migrates the package settings to use the Unity Settings Manager system rather than a ScriptableObject in the assets folder.